Started on docs, fixes to larch start script, split gui to pyqt folder
The start script should now work in desktop files, etc., so long as
gksu or kdesu is installed.

There is now a pyqt folder in modules containing the code specific to
the pyqt gui. There is the main, non larch specific, code in guibuild.py
and the larch specific starter code in larchgui.py.