Preparation for new hooks in rc. scripts.
The code to work with the new hooks should now be in place. To retain
compatibility with the pre-hook initscripts, I am replacing
rc.sysinit and rc.shutdown in inittab. My replacement script tests
for the function run_hook in rc.d/functions. If present it just
runs the normal rc. script, otherwise it adds the hook code and hooks
itself by editing te rc.script (to a temporary location).