Added some buttons (quit, cancel, help) and download script with progress
The download script is not yet used ... it could be patched in just for
building, but only if curl is available.

The cancel button doesn't do anything yet.

I will probably want to offer some configuration possibility for the
documentation browser.

I also added a hack for what I assume is a bug in the current version
of (py)qt - enabling of widgets with disabled parents works unexpectedly.