Fixed misnaming of new buttons in last commit, added download progress logging
In the last commit I had given the new buttons &-names, but their signal
handlers should not be backgrounded, so I removed the '&'s.

Using the script buildscripts/pacman-dl as an 'XferCommand' in pacman.conf
should allow download progress to appear in the log display. I added it
as the config-option 'dl_progress', which can be set on the Installation
Tweaks page.