Tweaks to larch-rebuild, remove usb2bootiso
larch-rebuild should now be complete (except for bug-fixes). It only
merges in the overlays and rebuilds system.sqf and mods.sqf, anything
further should be handled by some other script.

The features of usb2bootiso are now included in larch, so it is no
longer needed (it hadn't been updated for larch-7 anyway).