Added ldm - Larch Display Manager
First commit of this project. It's supposed to be a 'simple' replacement
for xdm, gdm, kdm. Experimental!

Being based on pyqt it's not really 'light', but it is simple in that there
is not much to configure (unless you include the styling here - which is
only present in rudimentary form, using qt style sheets).

Multiple xorg sessions are supported, as is autologin at start-up.