xkmap: Tiny tweak
Note that it is probably a mistake to use the hal configuration, as this
will soon be removed from xorg. Should I go back to xorg.conf? Or wait for
the hal replacement?

Support for ldm may also be misguided, as it is not at all clear that ldm
is really usable.