larch: This is a rather sudden and very extensive update
Basically the gui was drastically reworked, to use the uipi package,
as was the back-end, which now uses pexpect to run subprocesses.
This has probably introduced quite a few bugs, so testing is needed.

The documentation was updated and extended.

A fairly complete, but pretty rough German translation of the ui is
now in place as an i18n example.