General updates concerning new initscripts (with hooks), mkinitcpio, KMS ...
The code needed to patch in initscript hooks has been removed, because the
Arch package now includes the hooks.
Change to inittab handling partly connected with the above. Now
inittab.larchsave will only be created if the version for the live system
is supplied (in the profile) as inittab.larch.
Some profile updates concerning xorg/kernel/kms changes, also included
agp_intel and i915 modules in the default mkinitcpio.conf so that intel
graphic chips don't behave badly.
The installation clean-up script is now called 'delarch' (it was 'larch0').